examples/mandleweb/pyproject.toml
[tool.poetry]
name = "mandleweb"
version = "0.1.0"
description = ""
authors = ["Olivier <olivier@carbonifer.io>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
flask = "^3.0.3"
cython = "^3.0.11"
numpy = "^2.1.1"
matplotlib = "^3.9.2"
requests = "^2.32.3"
pypng = "^0.20220715.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
